feat: website-file-breakdown-for-current-selection-003 - Extract work selection package
