+++ ui/biome.jsonc
@@ -80,11 +80,11 @@
         "src/features/current-selection/components/current-selection-widget.provider-session-selection.test.tsx",
         "src/features/current-selection/components/current-selection-widget.save.test.tsx",
         "src/features/current-selection/components/current-selection-widget.test.tsx",
-        "src/features/current-selection/components/execution-details.test.tsx",
+        "src/features/current-selection/work-selection/components/execution-details.test.tsx",
         "src/features/current-selection/components/provider-session-attempts.test.tsx",
         "src/features/current-selection/components/provider-session-detail-panel.test.tsx",
         "src/features/current-selection/components/state-node-detail.test.tsx",
-        "src/features/current-selection/components/work-item-card.test.tsx",
+        "src/features/current-selection/work-selection/components/work-item-card.test.tsx",
         "src/features/current-selection/workstation-selection/components/workstation-detail-card.editable-configuration.test.tsx",
         "src/features/current-selection/workstation-selection/components/workstation-detail-card.test.tsx",
         "src/features/current-selection/workstation-selection/components/workstation-prompt-monaco.test.ts",
@@ -94,7 +94,7 @@
         "src/features/current-selection/hooks/useCurrentSelection.test.tsx",
         "src/features/current-selection/workstation-selection/messages/workstation-detail.test.ts",
         "src/features/current-selection/selected-work-dispatch-history-helpers.test.ts",
-        "src/features/current-selection/state/executionDetails.test.ts",
+        "src/features/current-selection/work-selection/state/executionDetails.test.ts",
         "src/features/current-selection/useCurrentSelection.request-helpers.test.ts",
         "src/features/current-selection/useCurrentSelection.selection-helpers.test.ts",
         "src/features/dashboard/hooks/useDashboardSnapshot.test.tsx",
