+++ ui/src/features/current-selection/components/detail-card-types.ts
@@ -1,2 +1,3 @@
 export * from "../base/components/detail-card-types";
 export * from "../workstation-selection/lib/detail-card-types";
+export * from "../work-selection/lib/detail-card-types";
