+++ ui/src/features/current-selection/public/index.ts
@@ -1,7 +1,8 @@
 export * from "../components/current-selection-widget";
 export * from "../hooks/useCurrentSelection";
 export * from "../hooks/useCurrentSelectionDetails";
-export * from "../hooks/useSelectedProviderSessionState";
+export { useSelectedProviderSessionState } from "../work-selection/public";
+export type { SelectedProviderSessionState } from "../work-selection/public";
 export type {
   DashboardSelection,
   DashboardWorkItemSelection,
