+++ ui/src/features/current-selection/work-selection/components/execution-details.test.tsx
@@ -1,8 +1,8 @@
 import { fireEvent, render, screen, within } from "@testing-library/react";
 import { describe, expect, it, vi } from "vitest";
-import { formatLocalDateTime } from "../../../components/ui/formatters";
-import { CurrentSelectionLocaleProvider } from "./current-selection-locale";
-import { DETAIL_CARD_NOW, inferenceAttempt } from "./detail-card-test-helpers";
+import { formatLocalDateTime } from "../../../../components/ui/formatters";
+import { CurrentSelectionLocaleProvider } from "../../components/current-selection-locale";
+import { DETAIL_CARD_NOW, inferenceAttempt } from "../../components/detail-card-test-helpers";
 import {
   ExecutionDetailsSection,
   InferenceAttemptsSection,
