+++ ui/src/features/current-selection/work-selection/lib/selected-work-relationship-graph.test.ts
@@ -3,7 +3,7 @@ import { describe, expect, it } from "vitest";
 import type {
   DashboardSnapshot,
   DashboardWorkItemRef,
-} from "../../../api/dashboard/types";
+} from "../../../../api/dashboard/types";
 import { buildSelectedWorkRelationshipGraph } from "./selected-work-relationship-graph";
 
 const selectedWorkItem: DashboardWorkItemRef = {
