+++ ui/src/features/current-selection/work-selection/public/index.ts
@@ -1,0 +1,38 @@
+export { WorkItemDetailCard } from "../components/work-item-card";
+export { TerminalWorkSummaryCard } from "../components/terminal-work-summary-detail";
+export {
+  ExecutionDetailsSection,
+  InferenceAttemptsSection,
+} from "../components/execution-details";
+export { InferenceAttemptCard } from "../components/inference-attempt";
+export { WorkItemPayloadList } from "../components/work-item-payload-details";
+export { WorkRelationshipsSection } from "../components/work-item-relationship-graph";
+
+export { useSelectedProviderSessionState } from "../hooks/useSelectedProviderSessionState";
+export type { SelectedProviderSessionState } from "../hooks/useSelectedProviderSessionState";
+
+export {
+  selectWorkItemExecutionDetails,
+} from "../state/executionDetails";
+export type {
+  SelectedWorkItemExecutionDetails,
+  SelectWorkItemExecutionDetailsInput,
+} from "../state/executionDetails";
+
+export {
+  buildSelectedWorkRelationshipGraph,
+} from "../lib/selected-work-relationship-graph";
+export type {
+  SelectedWorkRelationshipGraph,
+  SelectedWorkRelationshipNode,
+  SelectedWorkRelationshipEdge,
+  SelectedWorkRelationshipRole,
+} from "../lib/selected-work-relationship-graph";
+
+export type {
+  ExecutionDetailsSectionProps,
+  InferenceAttemptCardProps,
+  InferenceAttemptsSectionProps,
+  WorkItemDetailCardProps,
+  TerminalWorkSummaryCardProps,
+} from "../lib/detail-card-types";
